jenkins: make resourceRef required in JenkinsApi#buildProject
This method is only called in one place outside tests, and the
resourceRef is guaranteed to be set there, so we can tighten this
typing up a bit to make the value mandatory.

This class is not exported by the package, so this can be shipped as
a patch change.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>